Move Gemini calls to backend
