+++ src/app/generate-program/page.tsx
@@ -4,28 +4,12 @@ import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
 import { useUser } from "@clerk/nextjs";
 import { useEffect, useRef, useState } from "react";
-import { GoogleGenerativeAI } from "@google/generative-ai";
-
-const genAI = new GoogleGenerativeAI(process.env.GEMINI_API_KEY!);
 
 interface ChatMessage {
   role: "assistant" | "user";
   content: string;
 }
 
-const systemPrompt = `
-  You are a helpful fitness assistant. Your job is to collect and validate the following fields:
-  - Age (13–100)
-  - Height (in cm, 100–250)
-  - Weight (in kg or lbs, 30–200kg or 66–440lbs)
-  - Fitness goal (e.g., lose fat, build muscle)
-  - Workout days (e.g., monday, wednesday, friday, etc.)
-  - Fitness level (e.g.,beginner, intermediate, advanced)
-  - Injuries (e.g., teating ACL, wrist injuries, etc.)
-  - Diet preference (e.g., vegetarian, keto, none)
-  Ask one question at a time. If the answer is missing or invalid, ask again.
-  Once all valid values are collected, say: "Thank you for the information, kindly click Generate Plan!"
-`;  
 
 const GenerateProgramPage = () => {
   const { user } = useUser();
@@ -48,14 +32,6 @@ const GenerateProgramPage = () => {
   const messageContainerRef = useRef<HTMLDivElement>(null);
   const inputRef = useRef<HTMLInputElement>(null);
 
-  const model = genAI.getGenerativeModel({
-    model: "gemini-2.0-flash-001",
-    generationConfig: {
-      temperature: 0.4, // lower temperature for more predictable outputs
-      topP: 0.9,
-      responseMimeType: "application/json",
-    },
-  });
 
   {/* AUTO SCROLL */}
   useEffect(() => {
@@ -95,25 +71,15 @@ const handleSend = async (userMessage: string) => {
   };
 
 const sendToGemini = async (conversation: ChatMessage[]) => {
-  const chatHistory = conversation.map((msg) => ({
-    role: msg.role === "assistant" ? "model" : "user", 
-    parts: [{ text: msg.content }],
-  }));
-
   try {
-    const chat = await model.startChat({
-      history: [
-        {
-          role: "system",
-          parts: [{ text: systemPrompt }],
-        },
-        ...chatHistory,
-      ],
+    const res = await fetch("/api/gemini", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({ messages: conversation }),
     });
-
-    const result = await chat.sendMessage(conversation[conversation.length - 1].content);
-    const reply = await result.response.text();
-    return reply;
+    if (!res.ok) throw new Error("Request failed");
+    const data = await res.json();
+    return data.reply as string;
   } catch (err) {
     console.error("Gemini Error:", err);
     return "Sorry, I had trouble generating a response.";
